fix:运营采购管理功能优化

+++ src/views/business/shopGoodsOperate/js/shopGoodsOperateMethod.js
@@ -1,6 +1,7 @@
 import { deleteApi, getOneApi, updateApi, addApi, pageMapApi } from '@/api/business/shopGoodsOperateApi.js'
 // import { getAllApplication } from '@/api/business/applicationApi.js'
 import { postSupplierListApi } from '@/api/business/supplierApi.js'
+import $Big from '@/libs/big.js'
 
 // 初始化方法
 const initMethods = {
@@ -19,6 +20,9 @@ const dataMethods = {
   // 重置表单数据
   resetCreateForm() {
     this.createFormData = {
+      boxVolume: '',
+      boxQuantity: '',
+      suplierGuid: '',
       shopGoodsGuid: '',
       shopName: '',
       clientId: '',
@@ -121,27 +125,53 @@ const handleMethods = {
     this.submitCreate(formName)
   },
   // 点击编辑按钮事件
-  handleUpdate(row) {
+  handleUpdate() {
+    const selectionDatas = this.$refs.vxeTableRef.selection
+    if (!selectionDatas || selectionDatas.length !== 1) {
+      this.$message.warning('请选择一条数据')
+    } else {
+      const row = selectionDatas[0]
+      this.getOne(row.guid).then(response => {
+        if (this.filterFormData.statusCode === '待下单') {
+          this.createOrderFormData = response.data
+          this.showOrderComponent = true
+          this.dialogStatus = 'update'
+        }
+        if (this.filterFormData.statusCode === '已下单') {
+          this.createFormData = response.data
+          this.dialogFormVisible = true
+          this.dialogStatus = 'update'
+        }
+      })
+    }
+  },
+  // 下单操作
+  handleOrder(row) {
     this.getOne(row.guid).then(response => {
       this.createFormData = response.data
-      if (response.data.fileList) {
-        this.fileList = response.data.fileList
-      }
+      this.createFormData.statusCode = '已下单'
+      this.getSupplierData(row.goodsGuid)
       this.dialogFormVisible = true
       this.dialogStatus = 'update'
     })
   },
   // 删除信息
-  handleRemove(row) {
-    this.$confirm('此操作将删除数据, 是否继续?', '提示', {
-      confirmButtonText: '确定',
-      cancelButtonText: '取消',
-      type: 'warning'
-    }).then(() => {
-      this.handleHttpMethod(deleteApi(row.guid), true, '正在删除中', true, '删除成功').then(res => {
-        this.pageList()
+  handleRemove() {
+    const selectionDatas = this.$refs.vxeTableRef.selection
+    if (!selectionDatas || selectionDatas.length !== 1) {
+      this.$message.warning('请选择一条数据')
+    } else {
+      const row = selectionDatas[0]
+      this.$confirm('此操作将删除数据, 是否继续?', '提示', {
+        confirmButtonText: '确定',
+        cancelButtonText: '取消',
+        type: 'warning'
+      }).then(() => {
+        this.handleHttpMethod(deleteApi(row.guid), true, '正在删除中', true, '删除成功').then(res => {
+          this.pageList()
+        })
       })
-    })
+    }
   },
   // 过滤方法
   handleFilter(params) {
@@ -198,6 +228,24 @@ const handleMethods = {
     this.handleHttpMethod(postSupplierListApi({ goodsGuid: goodsGuid }), true, '请求中...').then(res => {
       this.supplierData = res.data
     })
+  },
+  // 新增采购订单
+  handleCreateOrder() {
+    this.showOrderComponent = true
+    this.dialogStatus = 'create'
+    this.createOrderFormData = {}
+  },
+  supplierChange(param) {
+    if (param) {
+      this.supplierData.forEach(item => {
+        if (item.supplierGuid === param.value) {
+          this.createFormData.boxVolume = item.boxVolume
+          this.createFormData.boxQuantity = item.boxQuantity
+          this.createFormData.totalBox = new $Big(this.createFormData.purNumber || 0).div(item.boxQuantity).toFixed(1).toString()
+          this.createFormData.purVolume = new $Big(item.boxVolume || 0).times(this.createFormData.totalBox || 0).toFixed(2).toString()
+        }
+      })
+    }
   }
 }
